pyproject.toml
[project]
name = "unified-icc"
version = "0.1.0"
description = "Platform-agnostic gateway for managing AI coding agents via tmux"
readme = "README.md"
authors = [
    { name = "Agony5757", email = "chenzhaoyun94@gmail.com" },
]
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "libtmux>=0.50.0",
    "pyte>=0.8.2",
    "Pillow>=10.0.0",
    "aiofiles>=24.0.0",
    "structlog>=24.0.0",
    "pathspec>=0.12",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "typer[all]>=0.15.0",
    "rich>=13.0.0",
    "websockets>=13.0,<14.0",
    "pyyaml>=6.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
]

[project.scripts]
unified-icc = "unified_icc.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "pytest-timeout>=2.3",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]
docs = [
    "sphinx>=8.0",
    "sphinx-book-theme>=0.4.0",
    "myst-parser>=4.0",
    "sphinx-autodoc-typehints>=2.0",
]
server = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=13.0",
]
dev-server = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=13.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0",
    "pytest-timeout>=2.3",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/unified_icc"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "ARG", "G", "BLE", "SIM", "N"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ARG"]
"src/**" = ["ARG002", "PLR2004"]

[tool.pyright]
pythonVersion = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
timeout = 30
